+++ src/components/common/Slide.jsx
